Reduce code duplication of internal database fields and addition of "plainParams" feature.
In the past, at crashreport.coffee, code had to be duplicated for each
internal parameter added (like version and product). Now, those
parameters are handled as if they were custom params given from the
configuration.

At config.coffee params from user configuration (.yaml file) are combined
with the ones application must have, which are defined at combineConfigList
function calls.

Each file param entry can be an string or a json object containing name
and downloadAs parameters. When configuration is read and mixed with the
internal params, all file entries are converted to the json format.

For each param, a referencing database table is created, on the other
hand, plainParams are just a plain database field that is not queriable
like params.

PlainParams feature has been mirrored to symfile.coffe for coherence.